requirements.txt
altgraph==0.17
certifi==2020.6.20
chardet==3.0.4
dbus-python==1.2.16
docutils==0.16
idna==2.10
lockfile==0.12.2
notify2==0.3.1
Pillow==7.2.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
PyMsgBox==1.0.8
requests==2.24.0
simpleaudio==1.0.4
urllib3==1.25.10
